nvim: fix install of color scheme symlink
This is unique, we can't create a symlink from
~/.config/nvim/colors/molokai.vim back to this project because it is a
symlink itself so this doesn't correctly resolve. So let's just copy in
the symlink instead (for lack of a better option).

+++ install.sh
@@ -18,8 +18,13 @@ for srcfile in $(find $rootdir -type l -or -type f -not -path "${rootdir}/.git*"
         mv -i "$dstfile" "${dstfile}.orig"
     fi
     srcfile="$(realpath_alt $srcfile)"
-    echo "creating symlink $dstfile -> $srcfile"
     mkdir -p "$(dirname ${dstfile})"
-    ln -fs "$srcfile" "$dstfile"
+    if [ ! -L $srcfile ]; then
+        echo "creating symlink $dstfile -> $srcfile"
+        ln -fs "$srcfile" "$dstfile"
+    else
+        echo "copying symlink $dstfile -> $srcfile"
+        cp -P "$srcfile" "$dstfile"
+    fi
 done
